fix 删除文章后,tag存在的问题

+++ in-blog-core/src/main/java/com/ijson/blog/service/impl/PostServiceImpl.java
@@ -294,17 +294,18 @@ public class PostServiceImpl implements PostService {
         postDao.delete(postId);
         draftDao.removeDraft(postId);
         List<String> topicIds = entity.getTopicId();
-        //遍历所有tags,如果postCount-1 <=0,则需要将tag删除
-        topicIds.forEach(topicId -> {
-            TopicEntity topic = topicDao.find(topicId);
-            long lastCount = topic.getPostCount() - 1;
-            if (lastCount <= 0) {
-                topicDao.delete(topic.getId());
-            } else {
-                topicDao.subtract(topic);
-            }
-        });
-
+        if(CollectionUtils.isNotEmpty(topicIds)){
+            //遍历所有tags,如果postCount-1 <=0,则需要将tag删除
+            topicIds.forEach(topicId -> {
+                TopicEntity topic = topicDao.find(topicId);
+                long lastCount = topic.getPostCount() - 1;
+                if (lastCount <= 0) {
+                    topicDao.delete(topic.getId());
+                } else {
+                    topicDao.subtract(topic);
+                }
+            });
+        }
         return null;
     }
